requirements.txt
Flask==2.3.2
Jinja2==3.1.2
psycopg2-binary==2.9.6
Werkzeug==2.3.6
gunicorn==21.2.0
python-dotenv==1.0.0